success send item with real data

+++ app/src/main/java/stta/gabriel/ta_gabriel/model/ItemLaporan.kt
@@ -10,10 +10,10 @@ data class ItemLaporan(
     val pelapor: String,
     var status: Int,
     val lokasi: Lokasi,
-    val head: Int,
+    val head: Long,
     val id_user: Int,
-    val tanggal_laporan:String
+    val tanggal_laporan: String
 ) :Parcelable
 {
-    constructor() : this("", "", "", 0, Lokasi(0, 0), 0, 0, "")
+    constructor() : this("", "", "", 0, Lokasi(0.0, 0.0), 0, 0, "")
 }
